~ | Full parsing refacto
