+++ includes/error.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   error.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: cpoulain <cpoulain@student.42lehavre.fr>   +#+  +:+       +#+        */
+/*   By: cpoulain <cpoulain@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/28 13:06:01 by cpoulain          #+#    #+#             */
-/*   Updated: 2025/02/12 11:15:06 by cpoulain         ###   ########.fr       */
+/*   Updated: 2025/02/17 12:56:29 by cpoulain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,6 +24,8 @@
 
 # define ERR_TOO_MANY_ARGS		"%s%s:\t%s : Too many arguments.\n%s"
 # define ERR_INT_ERR_ALLOC		"%s%s:\tInternal alloc error.\n%s"
+# define ERR_PARSE_ERROR_PIPE	"%s%s:\tParse error near `|'\n%s"
+# define ERR_PARSE_ERROR_QUOTE	"%s%s:\tUnmatched quote.\n%s"
 # define ERR_ARG_MUST_BE_INT	"%s%s:\t%s\t: must be an int.\n%s"
 # define ERR_CMD_NOT_FOUND		"%s%s:\tCommand not found : %s\n%s"
 # define ERR_COULD_NOT_OPEN_IN	"%s%s:\tCould not open %s.\n%s"
