+++ includes/parser.h
@@ -6,18 +6,15 @@
 /*   By: cpoulain <cpoulain@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/11 09:13:39 by jcheron           #+#    #+#             */
-/*   Updated: 2025/02/13 17:42:58 by cpoulain         ###   ########.fr       */
+/*   Updated: 2025/02/17 13:24:51 by cpoulain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef PARSER_H
 # define PARSER_H
 
-char			*args_to_string(
-					int ac,
-					char **args);
-
 t_list			*tokenize(
+					t_minishell_ctx *ctx,
 					const char *input);
 
 void			process_token(
@@ -44,33 +41,16 @@ int				handle_word_tokens(
 					size_t *i,
 					int *in_quotes);
 
-void			handle_unmatched_quotes(
-					t_list **tokens);
-
-void			handle_pipe(
-					t_cmd ***cmds,
-					size_t *cmd_count,
-					size_t *args_count);
-
 void			add_redirection(
 					t_list **tmp,
 					t_list **redir_list,
 					int type);
-
-void			add_argument(
-					t_cmd **current_cmd,
-					t_token **tmp,
-					size_t *args_count);
 
 t_list			*new_redir(
 					char *file,
 					t_redir_type type);
 
 t_cmd			*new_cmd(void);
-
-void			add_token(
-					t_list **head,
-					t_token *new);
 
 t_list			*new_token(
 					t_redir_type type,
@@ -79,21 +59,14 @@ t_list			*new_token(
 void			free_token(
 					void *token);
 
-int				handle_redir_in(
-					t_list **tokens,
-					const char *input,
-					size_t *i);
-
-int				handle_redir_out(
-					t_list **tokens,
-					const char *input,
-					size_t *i);
-
 void			add_cmd(
 					t_cmd ***cmds,
 					size_t *cmd_count);
 
-void			trim_quotes(
-					char **str);
+void			handle_parse_error(
+					t_minishell_ctx *ctx,
+					const char *error,
+					t_list **tokens
+					);
 
 #endif
